requirements.txt
# ShadowForge AI OS v1.0 - AI Operating System Dependencies
# Complete AI-controlled operating system with business intelligence

# Core Framework
crewai>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8

# AI Models & Inference
ollama>=0.1.7
openai>=1.10.0
anthropic>=0.8.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Quantum Computing & Advanced Math
qiskit>=0.45.0
numpy>=1.24.0
scipy>=1.10.0
sympy>=1.12.0

# DeFi & Blockchain Integration
web3>=6.0.0
eth-account>=0.8.0
requests>=2.31.0

# Data Processing & Analysis
pandas>=2.0.0
polars>=0.20.0
dask>=2023.0.0
pyarrow>=12.0.0

# Machine Learning & Neural Networks
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.2.0

# Computer Vision & Media
opencv-python>=4.8.0
pillow>=10.0.0
moviepy>=1.0.3
imageio>=2.31.0

# Natural Language Processing
spacy>=3.6.0
nltk>=3.8.0
textblob>=0.17.0
gensim>=4.3.0

# Web & API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0.0
aiohttp>=3.8.0
httpx>=0.25.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
motor>=3.3.0  # MongoDB async driver
asyncpg>=0.28.0  # PostgreSQL async driver

# Configuration & Environment
pydantic>=2.4.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.0

# Security & Cryptography
cryptography>=41.0.0
pycryptodome>=3.19.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.1.0
rich>=13.5.0
typer>=0.9.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Performance & Optimization
cython>=3.0.0
numba>=0.58.0
joblib>=1.3.0

# Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.13.0

# Real-time & Streaming
kafka-python>=2.0.0
pika>=1.3.0  # RabbitMQ
celery>=5.3.0

# Social Media & Content APIs
tweepy>=4.14.0
youtube-dl>=2021.12.17
instaloader>=4.9.0

# Financial Data & Trading
yfinance>=0.2.18
ccxt>=4.0.0
ta-lib>=0.4.28
backtrader>=1.9.76

# System & OS Integration
psutil>=5.9.0
watchdog>=3.0.0
schedule>=1.2.0

# Development & Debugging
ipython>=8.15.0
jupyter>=1.0.0
memory-profiler>=0.61.0
line-profiler>=4.1.0